+++ atelier/test.py
@@ -75,7 +75,8 @@ class TestCase(unittest.TestCase, SubProcessParent):
                     ok = True
                     args = [sys.executable]
                     args += ["-m"]
-                    args += ["atelier.doctest_utf8"]
+                    args += ["doctest"]
+                    # args += ["atelier.doctest_utf8"]
                     args += [fn]
                     self.run_subprocess(args, **kw)
                 if not ok: